requirements.txt
# FastAPI - Modern web framework for building APIs
fastapi==0.109.0

# Uvicorn - ASGI server to run FastAPI
uvicorn==0.27.0

# Python-multipart - Required for file uploads
python-multipart==0.0.6

# PyPDF2 - PDF text extraction library
PyPDF2==3.0.1

# Groq - Fast AI inference API (FREE!)
groq>=0.11.0

# Python-dotenv - Load environment variables from .env file
python-dotenv==1.0.0

# Pydantic - Data validation (comes with FastAPI but explicit is better)
pydantic==2.6.0
